Refs #18 - Parameter name collision when there Dapper expands arrays with 0 elements after 10th element
Previously we were naming arrays like @parray1, which Dapper expands to @parray11, @parray12, @parray13, etc.
However if there's another array at position 11 it would be named @parray11, and Dapper was supposed to expand it to @parray111, @parray112, etc. However if there are no elements it renders as @parray11, conflicting with the other array.

Now arrays are named like @p1array, so there are no more conflicts.